refactor: drop legacy style engine
The only left usage of legacy style engine was collapsed instance logic.
Here inlined necessary code and got rid of the rest.
