+++ apps/builder/package.json
@@ -124,7 +124,6 @@
   "devDependencies": {
     "@jest/globals": "^29.7.0",
     "@remix-run/dev": "^2.12.1",
-    "@testing-library/react-hooks": "^8.0.1",
     "@types/cookie": "^0.6.0",
     "@types/debug": "^4.1.12",
     "@types/dom-navigation": "^1.0.4",
